Lab_4/requirements.txt
numpy
pandas
python-docx
statistics

